Changed prod config to avoid extra databases
Using solid cache, queue, etc. will require adding separate databases for each of them to set them up optimally. We don't need all that extra complexity and cost just for a simple application. Set the cache store to :memory_store for now. We can revisit if we actually want to make this into a larger app.